--ArtLink classes into separate header; links get owning Managed AO
